+++ api-module/src/main/java/com/romanticpipe/reviewcanvas/domain/review/application/usecase/ReviewUseCaseImpl.java
@@ -17,12 +17,8 @@ import com.romanticpipe.reviewcanvas.domain.review.application.usecase.response.
 import com.romanticpipe.reviewcanvas.dto.PageResponse;
 import com.romanticpipe.reviewcanvas.dto.PageableRequest;
 import com.romanticpipe.reviewcanvas.enumeration.ReviewFilter;
-import com.romanticpipe.reviewcanvas.service.ProductCreator;
-import com.romanticpipe.reviewcanvas.service.ProductReader;
-import com.romanticpipe.reviewcanvas.service.ProductValidator;
-import com.romanticpipe.reviewcanvas.service.ReviewCreator;
-import com.romanticpipe.reviewcanvas.service.ReviewReader;
-import com.romanticpipe.reviewcanvas.service.ReviewValidator;
+import com.romanticpipe.reviewcanvas.service.ProductService;
+import com.romanticpipe.reviewcanvas.service.ReviewService;
 
 import lombok.RequiredArgsConstructor;
 
@@ -31,12 +27,8 @@ import lombok.RequiredArgsConstructor;
 class ReviewUseCaseImpl implements ReviewUseCase {
 
 	private final ShopAdminService shopAdminService;
-	private final ProductCreator productCreator;
-	private final ProductValidator productValidator;
-	private final ProductReader productReader;
-	private final ReviewReader reviewReader;
-	private final ReviewCreator reviewCreator;
-	private final ReviewValidator reviewValidator;
+	private final ProductService productService;
+	private final ReviewService reviewService;
 	private final Cafe24ProductClient cafe24ProductClient;
 	private final TransactionTemplate writeTransactionTemplate;
 
@@ -44,30 +36,30 @@ class ReviewUseCaseImpl implements ReviewUseCase {
 	public PageResponse<GetReviewForUserResponse> getReviewsForUser(String mallId, Long productNo,
 		PageableRequest pageableRequest,
 		ReviewFilter filter) {
-		Product product = productReader.findProduct(mallId, productNo)
+		Product product = productService.findProduct(mallId, productNo)
 			.orElseGet(() -> createProduct(mallId, productNo));
 
-		return reviewReader.findAllByProductId(product.getId(), pageableRequest, filter)
+		return reviewService.findAllByProductId(product.getId(), pageableRequest, filter)
 			.map(GetReviewForUserResponse::from);
 	}
 
 	@Override
 	@Transactional(readOnly = true)
 	public GetReviewForUserResponse getReviewForUser(Long reviewId) {
-		return GetReviewForUserResponse.from(reviewValidator.validateUserInfoById(reviewId));
+		return GetReviewForUserResponse.from(reviewService.validateUserInfoById(reviewId));
 	}
 
 	@Override
 	@Transactional(readOnly = true)
 	public PageResponse<GetReviewResponse> getReviewsByUserId(String userId, PageableRequest pageableRequest) {
-		return reviewReader.findByUserId(userId, pageableRequest)
+		return reviewService.findByUserId(userId, pageableRequest)
 			.map(GetReviewResponse::from);
 	}
 
 	@Override
 	@Transactional
 	public void updateReview(long reviewId, UpdateReviewRequest updateReviewRequest) {
-		Review review = reviewValidator.validById(reviewId);
+		Review review = reviewService.validById(reviewId);
 		review.setScore(updateReviewRequest.score());
 		review.setContent(updateReviewRequest.content());
 	}
@@ -75,7 +67,7 @@ class ReviewUseCaseImpl implements ReviewUseCase {
 	@Override
 	@Transactional
 	public void createReview(String productId, CreateReviewRequest createReviewRequest) {
-		Product product = productValidator.validByProductId(productId);
+		Product product = productService.validByProductId(productId);
 		// TODO: product가  mallId와 productNo로 상품을 생성하는 로직을 추가하도록 변경해야 함.
 		//		Product product = productReader.findByMallIdAndProductNo(mallId, productNo)
 		//			.orElseGet(() -> createProduct(mallId, productNo));
@@ -99,7 +91,7 @@ class ReviewUseCaseImpl implements ReviewUseCase {
 		return writeTransactionTemplate.execute(status -> {
 			ShopAdmin shopAdmin = shopAdminService.validByMallId(mallId);
 			Product product = new Product(productNo, cafe24ProductDto.product().productName(), shopAdmin.getId());
-			return productCreator.save(product);
+			return productService.save(product);
 		});
 	}
 }
