--- domain-module/review/src/main/java/com/romanticpipe/reviewcanvas/service/ProductCreator.java
@@ -1,17 +1,0 @@
-package com.romanticpipe.reviewcanvas.service;
-
-import com.romanticpipe.reviewcanvas.domain.Product;
-import com.romanticpipe.reviewcanvas.repository.ProductRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class ProductCreator {
-
-	private final ProductRepository productRepository;
-
-	public Product save(Product product) {
-		return productRepository.save(product);
-	}
-}
