+++ domain-module/review/src/main/java/com/romanticpipe/reviewcanvas/service/ProductService.java
@@ -1,17 +1,34 @@
 package com.romanticpipe.reviewcanvas.service;
 
+import java.util.List;
+import java.util.Optional;
+
+import org.springframework.stereotype.Service;
+
 import com.romanticpipe.reviewcanvas.domain.Product;
 import com.romanticpipe.reviewcanvas.exception.BusinessException;
 import com.romanticpipe.reviewcanvas.exception.ProductErrorCode;
 import com.romanticpipe.reviewcanvas.repository.ProductRepository;
+
 import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
 
 @Service
 @RequiredArgsConstructor
-public class ProductValidator {
+public class ProductService {
 
 	private final ProductRepository productRepository;
+
+	public Product save(Product product) {
+		return productRepository.save(product);
+	}
+
+	public Optional<Product> findProduct(String mallId, Long productNo) {
+		return productRepository.findByMallIdAndProductNo(mallId, productNo);
+	}
+
+	public List<Product> findProducts(Integer shopAdminId) {
+		return productRepository.findAllByShopAdminId(shopAdminId);
+	}
 
 	public Product validByProductId(String productId) {
 		return productRepository.findById(productId)
