--- domain-module/review/src/main/java/com/romanticpipe/reviewcanvas/service/ReviewCreator.java
@@ -1,17 +1,0 @@
-package com.romanticpipe.reviewcanvas.service;
-
-import com.romanticpipe.reviewcanvas.domain.Review;
-import com.romanticpipe.reviewcanvas.repository.ReviewRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class ReviewCreator {
-
-	private final ReviewRepository reviewRepository;
-
-	public void save(Review review) {
-		reviewRepository.save(review);
-	}
-}
