--- domain-module/review/src/main/java/com/romanticpipe/reviewcanvas/service/ReviewValidator.java
@@ -1,27 +1,0 @@
-package com.romanticpipe.reviewcanvas.service;
-
-import com.romanticpipe.reviewcanvas.domain.Review;
-import com.romanticpipe.reviewcanvas.dto.ReviewInfo;
-import com.romanticpipe.reviewcanvas.exception.BusinessException;
-import com.romanticpipe.reviewcanvas.exception.ReviewErrorCode;
-import com.romanticpipe.reviewcanvas.repository.ReviewRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class ReviewValidator {
-
-	private final ReviewRepository reviewRepository;
-
-	public Review validById(long reviewId) {
-		return reviewRepository.findById(reviewId)
-			.orElseThrow(() -> new BusinessException(ReviewErrorCode.REVIEW_NOT_FOUND));
-	}
-
-	public ReviewInfo validateUserInfoById(Long reviewId) {
-		return reviewRepository.findReviewInfoById(reviewId)
-			.orElseThrow(() -> new BusinessException(ReviewErrorCode.REVIEW_NOT_FOUND));
-	}
-
-}
